Updates to make this work after refactors
